Fixing seaching by transcation amount

+++ K9-Koinz/Pages/Transactions/Index.cshtml.cs
@@ -141,7 +141,7 @@ namespace K9_Koinz.Pages.Transactions
             }
 
             if (!string.IsNullOrWhiteSpace(searchString)) {
-                if (float.TryParse(searchString, out float value)) {
+                if (double.TryParse(searchString, out double value)) {
                     transactionsIQ = transactionsIQ.Where(trans => trans.Amount == value || trans.Amount == -1 * value);
                 } else if (searchString.ToLower() == "hidden") {
                     transactionsIQ = transactionsIQ.Where(trans => trans.IsSavingsSpending);
